ir.model.fields: implement the write(), allow custom fields to change
Implement the missing feature of being able to modify those custom
fields in the database, through the ORM interface. Now, using the
ir.model or ir.model.fields forms, the admin is able to alter properties
of custom fields on a live server (without server restart).

In the future, we may even remove the fuse for base fields and allow
them to change, too.