[FIX] ir.ui.view,ir.ui.view.custom: cleanup handling of customized views + add menu and views
Custom views are used when users reorganize their dashboards (drag and drop in web), but this should not be
special-cased for board.board model in the core. The web client will instead directly write to the ir.ui.view.custom
table, and the board module will override board.board.fields_view_get to take custom views into account.
The only responsibility of the core is to drop custom views when the original view is updated, to be sure there
are no outdated customizations that remain.