expression: fix function.search() when it does not restrict
Triggered by [Bug 702340]

When we try to search a composite expression like:
   [( 'ourfoo.subfoo', '=' , True) ]
.. where subfoo is a function field of the model ofourfoo field,
the framework will call subfoo.search([('subfoo','=',True)]).
The latter shall return a domain expression that will restrict
the search on 'ourfoo'.

When the returned expression is empty, meaning "match all", the algebra
on the parent expression would come up with an invalid  [..., '&', ( 1, '=', 1) ]
sequence.