[FIX] requirements: Don't complain on missing library
Module sms wizard try to import this library and complains with a warning
if not found. This makes runbot to turn brown, which for Travis status
means red, causing falses alarms by that.

As most of the 11.0 repositories uses OCB for building runbot instances,
this will greenify the PRs without adding the library repository per
repository.